+++ R/conditions.R
@@ -422,44 +422,44 @@ stop_names_must_be_unique <- function(locations) {
 #'
 #' @description
 #'
-#' * `stop_index_oob_position()` throws errors of class
-#'   `vctrs_error_index_oob_position` containing fields `i` and
+#' * `stop_index_oob_positions()` throws errors of class
+#'   `vctrs_error_index_oob_positions` containing fields `i` and
 #'   `size`.
 #'
-#' * `stop_index_oob_name()` throws errors of class
-#'   `vctrs_error_index_oob_name` containing fields `i` and `names`.
+#' * `stop_index_oob_names()` throws errors of class
+#'   `vctrs_error_index_oob_names` containing fields `i` and `names`.
 #'
-#' @param i For `stop_index_oob_position()`, a numeric vector of
-#'   positions. For `stop_index_oob_name()`, a character vector of
+#' @param i For `stop_index_oob_positions()`, a numeric vector of
+#'   positions. For `stop_index_oob_names()`, a character vector of
 #'   names. `i` may contain both out-of-bounds and within-bounds
 #'   elements, only the former are used to construct the error
 #'   message.
 #' @param size The length of the vector to subset from.
 #' @inheritParams rlang::abort
 #' @export
-stop_index_oob_position <- function(i, size, ..., .subclass = NULL) {
+stop_index_oob_positions <- function(i, size, ..., .subclass = NULL) {
   abort(
     "",
-    .subclass = c(.subclass, "vctrs_error_index_oob_position"),
+    .subclass = c(.subclass, "vctrs_error_index_oob_positions"),
     i = i,
     size = size,
     ...
   )
 }
-#' @rdname stop_index_oob_position
+#' @rdname stop_index_oob_positions
 #' @param names The names of the vector to subset from.
 #' @export
-stop_index_oob_name <- function(i, names, ..., .subclass = NULL) {
+stop_index_oob_names <- function(i, names, ..., .subclass = NULL) {
   abort(
     "",
-    .subclass = c(.subclass, "vctrs_error_index_oob_name"),
+    .subclass = c(.subclass, "vctrs_error_index_oob_names"),
     i = i,
     names = names,
     ...
   )
 }
 #' @export
-conditionMessage.vctrs_error_index_oob_position <- function(c) {
+conditionMessage.vctrs_error_index_oob_positions <- function(c) {
   if (!nzchar(c$message)) {
     i <- c$i
 
@@ -483,7 +483,7 @@ conditionMessage.vctrs_error_index_oob_position <- function(c) {
   NextMethod()
 }
 #' @export
-conditionMessage.vctrs_error_index_oob_name <- function(c) {
+conditionMessage.vctrs_error_index_oob_names <- function(c) {
   if (!nzchar(c$message)) {
     oob <- c$i[!c$i %in% c$names]
     oob_enum <- enumerate(glue::backtick(oob))